move RGBLight functionality into Light
