+++ test/devices_tests/light_test.py
@@ -65,6 +65,44 @@ class TestLight(unittest.TestCase):
         light = Light(xknx, "Diningroom.Light_1", group_address_switch="1/6/4")
         self.assertFalse(light.supports_color)
 
+    def test_supports_individual_color_true(self):
+        """Test supports_color true."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+        )
+        self.assertTrue(light.supports_color)
+
+    def test_supports_individual_color_false(self):
+        """Test supports_color false."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+        )
+        self.assertFalse(light.supports_color)
+
     #
     # TEST SUPPORT COLOR RGBW
     #
@@ -88,6 +126,52 @@ class TestLight(unittest.TestCase):
             "Diningroom.Light_1",
             group_address_switch="1/6/4",
             group_address_color="1/6/6",
+        )
+        self.assertFalse(light.supports_rgbw)
+
+    def test_supports_individual_rgbw_true(self):
+        """Test supports_rgbw true."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+            group_address_switch_white="1/1/13",
+            group_address_switch_state_white="1/1/14",
+            group_address_brightness_white="1/1/15",
+            group_address_brightness_state_white="1/1/16",
+        )
+        self.assertTrue(light.supports_rgbw)
+
+    def test_supports_individual_rgbw_false(self):
+        """Test supports_color false."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
         )
         self.assertFalse(light.supports_rgbw)
 
@@ -179,6 +263,73 @@ class TestLight(unittest.TestCase):
         self.assertEqual(len(telegrams), 6)
         self.assertListEqual(telegrams, test_telegrams)
 
+    def test_sync_individual_color(self):
+        """Test sync function / sending group reads to KNX bus. Testing with a Light without dimm functionality."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+            group_address_switch_white="1/1/13",
+            group_address_switch_state_white="1/1/14",
+            group_address_brightness_white="1/1/15",
+            group_address_brightness_state_white="1/1/16",
+        )
+        self.loop.run_until_complete(light.sync())
+
+        self.assertEqual(xknx.telegrams.qsize(), 8)
+
+        telegrams = [xknx.telegrams.get_nowait() for _ in range(8)]
+
+        test_telegrams = [
+            Telegram(
+                destination_address=GroupAddress("1/1/2"),
+                telegramtype=TelegramType.GROUP_READ,
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/4"),
+                telegramtype=TelegramType.GROUP_READ,
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/6"),
+                telegramtype=TelegramType.GROUP_READ,
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/8"),
+                telegramtype=TelegramType.GROUP_READ,
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/10"),
+                telegramtype=TelegramType.GROUP_READ,
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/12"),
+                telegramtype=TelegramType.GROUP_READ,
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/14"),
+                telegramtype=TelegramType.GROUP_READ,
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/16"),
+                telegramtype=TelegramType.GROUP_READ,
+            ),
+        ]
+
+        self.assertEqual(len(set(telegrams)), 8)
+        self.assertEqual(set(telegrams), set(test_telegrams))
+
     #
     # SYNC WITH STATE ADDRESS
     #
@@ -256,6 +407,51 @@ class TestLight(unittest.TestCase):
             ),
         )
 
+    def test_set_on_individual_color(self):
+        """Test switching on a Light."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+            group_address_switch_white="1/1/13",
+            group_address_switch_state_white="1/1/14",
+            group_address_brightness_white="1/1/15",
+            group_address_brightness_state_white="1/1/16",
+        )
+        self.loop.run_until_complete(light.set_on())
+        self.assertEqual(xknx.telegrams.qsize(), 4)
+
+        telegrams = [xknx.telegrams.get_nowait() for _ in range(4)]
+
+        test_telegrams = [
+            Telegram(
+                destination_address=GroupAddress("1/1/1"), payload=DPTBinary(True)
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/5"), payload=DPTBinary(True)
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/9"), payload=DPTBinary(True)
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/13"), payload=DPTBinary(True)
+            ),
+        ]
+        self.assertEqual(len(set(telegrams)), 4)
+        self.assertEqual(set(telegrams), set(test_telegrams))
+
     #
     # TEST SET OFF
     #
@@ -278,6 +474,51 @@ class TestLight(unittest.TestCase):
                 payload=GroupValueWrite(DPTBinary(0)),
             ),
         )
+
+    def test_set_off_individual_color(self):
+        """Test switching off a Light."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+            group_address_switch_white="1/1/13",
+            group_address_switch_state_white="1/1/14",
+            group_address_brightness_white="1/1/15",
+            group_address_brightness_state_white="1/1/16",
+        )
+        self.loop.run_until_complete(light.set_off())
+        self.assertEqual(xknx.telegrams.qsize(), 4)
+
+        telegrams = [xknx.telegrams.get_nowait() for _ in range(4)]
+
+        test_telegrams = [
+            Telegram(
+                destination_address=GroupAddress("1/1/1"), payload=DPTBinary(False)
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/5"), payload=DPTBinary(False)
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/9"), payload=DPTBinary(False)
+            ),
+            Telegram(
+                destination_address=GroupAddress("1/1/13"), payload=DPTBinary(False)
+            ),
+        ]
+        self.assertEqual(len(set(telegrams)), 4)
+        self.assertEqual(set(telegrams), set(test_telegrams))
 
     #
     # TEST SET BRIGHTNESS
@@ -351,6 +592,77 @@ class TestLight(unittest.TestCase):
                 "Colors not supported for device %s", "TestLight"
             )
 
+    def test_set_individual_color(self):
+        """Test setting the color of a Light."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+        )
+        self.loop.run_until_complete(light.set_color([23, 24, 25]))
+        self.assertEqual(xknx.telegrams.qsize(), 3)
+        telegrams = [xknx.telegrams.get_nowait() for _ in range(3)]
+
+        test_telegrams = [
+            Telegram(destination_address=GroupAddress("1/1/3"), payload=DPTArray(23)),
+            Telegram(destination_address=GroupAddress("1/1/7"), payload=DPTArray(24)),
+            Telegram(destination_address=GroupAddress("1/1/11"), payload=DPTArray(25)),
+        ]
+
+        self.assertEqual(len(set(telegrams)), 3)
+        self.assertEqual(set(telegrams), set(test_telegrams))
+
+        self.loop.run_until_complete(
+            xknx.devices.process(
+                Telegram(
+                    destination_address=GroupAddress("1/1/4"), payload=DPTArray(23)
+                )
+            )
+        )
+        self.loop.run_until_complete(
+            xknx.devices.process(
+                Telegram(
+                    destination_address=GroupAddress("1/1/8"), payload=DPTArray(24)
+                )
+            )
+        )
+        self.loop.run_until_complete(
+            xknx.devices.process(
+                Telegram(
+                    destination_address=GroupAddress("1/1/12"), payload=DPTArray(25)
+                )
+            )
+        )
+        self.assertEqual(light.current_color, ([23, 24, 25], None))
+
+    def test_set_individual_color_not_possible(self):
+        """Test setting the color of a non light without color."""
+        # pylint: disable=invalid-name
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "TestLight",
+            group_address_switch_red="1/1/1",
+        )
+        with patch("logging.Logger.warning") as mock_warn:
+            self.loop.run_until_complete(light.set_color((23, 24, 25)))
+            self.assertEqual(xknx.telegrams.qsize(), 0)
+            mock_warn.assert_called_with(
+                "Colors not supported for device %s", "TestLight"
+            )
+
     #
     # TEST SET COLOR AS RGBW
     #
@@ -386,6 +698,101 @@ class TestLight(unittest.TestCase):
             name="TestLight",
             group_address_switch="1/2/3",
             group_address_color="1/2/4",
+        )
+        with patch("logging.Logger.warning") as mock_warn:
+            self.loop.run_until_complete(light.set_color((23, 24, 25), 26))
+
+            self.assertEqual(xknx.telegrams.qsize(), 0)
+            mock_warn.assert_called_with(
+                "RGBW not supported for device %s", "TestLight"
+            )
+
+    def test_set_individual_color_rgbw(self):
+        """Test setting RGBW value of a Light."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+            group_address_switch_white="1/1/13",
+            group_address_switch_state_white="1/1/14",
+            group_address_brightness_white="1/1/15",
+            group_address_brightness_state_white="1/1/16",
+        )
+        self.loop.run_until_complete(light.set_color([23, 24, 25], white=26))
+        self.assertEqual(xknx.telegrams.qsize(), 4)
+        telegrams = [xknx.telegrams.get_nowait() for _ in range(4)]
+
+        test_telegrams = [
+            Telegram(destination_address=GroupAddress("1/1/3"), payload=DPTArray(23)),
+            Telegram(destination_address=GroupAddress("1/1/7"), payload=DPTArray(24)),
+            Telegram(destination_address=GroupAddress("1/1/11"), payload=DPTArray(25)),
+            Telegram(destination_address=GroupAddress("1/1/15"), payload=DPTArray(26)),
+        ]
+
+        self.assertEqual(len(set(telegrams)), 4)
+        self.assertEqual(set(telegrams), set(test_telegrams))
+
+        self.loop.run_until_complete(
+            xknx.devices.process(
+                Telegram(
+                    destination_address=GroupAddress("1/1/4"), payload=DPTArray(23)
+                )
+            )
+        )
+        self.loop.run_until_complete(
+            xknx.devices.process(
+                Telegram(
+                    destination_address=GroupAddress("1/1/8"), payload=DPTArray(24)
+                )
+            )
+        )
+        self.loop.run_until_complete(
+            xknx.devices.process(
+                Telegram(
+                    destination_address=GroupAddress("1/1/12"), payload=DPTArray(25)
+                )
+            )
+        )
+        self.loop.run_until_complete(
+            xknx.devices.process(
+                Telegram(
+                    destination_address=GroupAddress("1/1/16"), payload=DPTArray(26)
+                )
+            )
+        )
+        self.assertEqual(light.current_color, ([23, 24, 25], 26))
+
+    def test_set_individual_color_rgbw_not_possible(self):
+        """Test setting RGBW value of a non light without color."""
+        # pylint: disable=invalid-name
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "TestLight",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
         )
         with patch("logging.Logger.warning") as mock_warn:
             self.loop.run_until_complete(light.set_color((23, 24, 25), 26))
@@ -500,6 +907,39 @@ class TestLight(unittest.TestCase):
         self.loop.run_until_complete(light.process(telegram))
         self.assertEqual(light.state, False)
 
+    def test_process_color_switch(self):
+        """Test process / reading telegrams from telegram queue. Test if switch position is processed correctly."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "TestLight",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+        )
+        self.assertEqual(light.state, None)
+
+        telegram = Telegram(
+            destination_address=GroupAddress("1/1/2"), payload=DPTBinary(True)
+        )
+        self.loop.run_until_complete(light.process(telegram))
+        self.assertEqual(light.state, True)
+
+        telegram = Telegram(
+            destination_address=GroupAddress("1/1/2"), payload=DPTBinary(False)
+        )
+        self.loop.run_until_complete(light.process(telegram))
+        self.assertEqual(light.state, False)
+
     def test_process_switch_callback(self):
         """Test process / reading telegrams from telegram queue. Test if callback is called."""
         # pylint: disable=no-self-use
@@ -595,6 +1035,37 @@ class TestLight(unittest.TestCase):
         self.loop.run_until_complete(light.process(telegram))
         self.assertEqual(light.current_color, ((23, 24, 25), None))
 
+    def test_process_individual_color(self):
+        """Test process / reading telegrams from telegram queue. Test if color is processed."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "TestLight",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+        )
+        self.assertEqual(light.current_color, (None, None))
+
+        telegrams = [
+            Telegram(destination_address=GroupAddress("1/1/4"), payload=DPTArray(42)),
+            Telegram(destination_address=GroupAddress("1/1/8"), payload=DPTArray(43)),
+            Telegram(destination_address=GroupAddress("1/1/12"), payload=DPTArray(44)),
+        ]
+
+        for telegram in telegrams:
+            self.loop.run_until_complete(light.process(telegram))
+        self.assertEqual(light.current_color, ([42, 43, 44], None))
+
     def test_process_color_rgbw(self):
         """Test process / reading telegrams from telegram queue. Test if RGBW is processed."""
         xknx = XKNX()
@@ -612,6 +1083,36 @@ class TestLight(unittest.TestCase):
         )
         self.loop.run_until_complete(light.process(telegram))
         self.assertEqual(light.current_color, ([23, 24, 25], 26))
+
+    def test_process_individual_color_rgbw(self):
+        """Test process / reading telegrams from telegram queue. Test if RGBW is processed."""
+        xknx = XKNX()
+        light = Light(
+            xknx,
+            "Diningroom.Light_1",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+            group_address_switch_white="1/1/13",
+            group_address_switch_state_white="1/1/14",
+            group_address_brightness_white="1/1/15",
+            group_address_brightness_state_white="1/1/16",
+        )
+        self.assertEqual(light.current_color, (None, None))
+        telegram = Telegram(
+            destination_address=GroupAddress("1/1/16"), payload=DPTArray(42)
+        )
+        self.loop.run_until_complete(light.process(telegram))
+        self.assertEqual(light.current_color, (None, 42))
 
     def test_process_tunable_white(self):
         """Test process / reading telegrams from telegram queue. Test if tunable white is processed."""
@@ -786,6 +1287,22 @@ class TestLight(unittest.TestCase):
             group_address_color_temperature_state="1/7/10",
             group_address_rgbw="1/7/11",
             group_address_rgbw_state="1/7/12",
+            group_address_switch_red="1/1/1",
+            group_address_switch_state_red="1/1/2",
+            group_address_brightness_red="1/1/3",
+            group_address_brightness_state_red="1/1/4",
+            group_address_switch_green="1/1/5",
+            group_address_switch_state_green="1/1/6",
+            group_address_brightness_green="1/1/7",
+            group_address_brightness_state_green="1/1/8",
+            group_address_switch_blue="1/1/9",
+            group_address_switch_state_blue="1/1/10",
+            group_address_brightness_blue="1/1/11",
+            group_address_brightness_state_blue="1/1/12",
+            group_address_switch_white="1/1/13",
+            group_address_switch_state_white="1/1/14",
+            group_address_brightness_white="1/1/15",
+            group_address_brightness_state_white="1/1/16",
         )
         self.assertTrue(light.has_group_address(GroupAddress("1/7/1")))
         self.assertTrue(light.has_group_address(GroupAddress("1/7/2")))
@@ -799,4 +1316,22 @@ class TestLight(unittest.TestCase):
         self.assertTrue(light.has_group_address(GroupAddress("1/7/10")))
         self.assertTrue(light.has_group_address(GroupAddress("1/7/11")))
         self.assertTrue(light.has_group_address(GroupAddress("1/7/12")))
+
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/1")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/2")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/3")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/4")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/5")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/6")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/7")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/8")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/9")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/10")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/11")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/12")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/13")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/14")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/15")))
+        self.assertTrue(light.has_group_address(GroupAddress("1/1/16")))
+
         self.assertFalse(light.has_group_address(GroupAddress("1/7/13")))
