+++ xknx/devices/light.py
@@ -12,7 +12,7 @@ It provides functionality for
 """
 from enum import Enum
 import logging
-from typing import Optional
+from typing import Optional, Tuple
 
 from xknx.remote_value import (
     RemoteValueColorRGB,
@@ -32,6 +32,57 @@ class ColorTempModes(Enum):
 
     absolute = "DPT-7.600"
     relative = "DPT-5.001"
+
+
+class _SwitchAndBrightness:
+    def __init__(
+        self,
+        xknx,
+        name,
+        feature_name: str,
+        group_address_switch=None,
+        group_address_switch_state=None,
+        group_address_brightness=None,
+        group_address_brightness_state=None,
+        after_update_cb=None,
+    ):
+        self.switch = RemoteValueSwitch(
+            xknx,
+            group_address_switch,
+            group_address_switch_state,
+            device_name=name,
+            feature_name=feature_name + "_state",
+            after_update_cb=after_update_cb,
+        )
+        self.brightness = RemoteValueScaling(
+            xknx,
+            group_address_brightness,
+            group_address_brightness_state,
+            device_name=name,
+            feature_name=feature_name + "_brightness",
+            after_update_cb=after_update_cb,
+            range_from=0,
+            range_to=255,
+        )
+
+    @property
+    def is_on(self):
+        """Return if light supports color."""
+        return self.switch.value
+
+    async def set_on(self):
+        """Switch light on."""
+        if self.switch.initialized:
+            await self.switch.on()
+
+    async def set_off(self):
+        """Switch light off."""
+        if self.switch.initialized:
+            await self.switch.off()
+
+    def __eq__(self, other):
+        """Compare for quality."""
+        return self.__dict__ == other.__dict__
 
 
 # pylint: disable=too-many-public-methods, too-many-instance-attributes
@@ -58,6 +109,22 @@ class Light(Device):
         group_address_tunable_white_state=None,
         group_address_color_temperature=None,
         group_address_color_temperature_state=None,
+        group_address_switch_red=None,
+        group_address_switch_state_red=None,
+        group_address_brightness_red=None,
+        group_address_brightness_state_red=None,
+        group_address_switch_green=None,
+        group_address_switch_state_green=None,
+        group_address_brightness_green=None,
+        group_address_brightness_state_green=None,
+        group_address_switch_blue=None,
+        group_address_switch_state_blue=None,
+        group_address_brightness_blue=None,
+        group_address_brightness_state_blue=None,
+        group_address_switch_white=None,
+        group_address_switch_state_white=None,
+        group_address_brightness_white=None,
+        group_address_brightness_state_white=None,
         min_kelvin=None,
         max_kelvin=None,
         device_updated_cb=None,
@@ -122,6 +189,50 @@ class Light(Device):
             after_update_cb=self.after_update,
         )
 
+        self.red = _SwitchAndBrightness(
+            xknx,
+            self.name,
+            "red",
+            group_address_switch_red,
+            group_address_switch_state_red,
+            group_address_brightness_red,
+            group_address_brightness_state_red,
+            self.after_update,
+        )
+
+        self.green = _SwitchAndBrightness(
+            xknx,
+            self.name,
+            "green",
+            group_address_switch_green,
+            group_address_switch_state_green,
+            group_address_brightness_green,
+            group_address_brightness_state_green,
+            self.after_update,
+        )
+
+        self.blue = _SwitchAndBrightness(
+            xknx,
+            self.name,
+            "blue",
+            group_address_switch_blue,
+            group_address_switch_state_blue,
+            group_address_brightness_blue,
+            group_address_brightness_state_blue,
+            self.after_update,
+        )
+
+        self.white = _SwitchAndBrightness(
+            xknx,
+            self.name,
+            "white",
+            group_address_switch_white,
+            group_address_switch_state_white,
+            group_address_brightness_white,
+            group_address_brightness_state_white,
+            self.after_update,
+        )
+
         self.min_kelvin = min_kelvin
         self.max_kelvin = max_kelvin
 
@@ -135,6 +246,9 @@ class Light(Device):
             self.tunable_white,
             self.color_temperature,
         )
+        for color in (self.red, self.green, self.blue, self.white):
+            yield color.switch
+            yield color.brightness
 
     @property
     def supports_brightness(self):
@@ -144,12 +258,19 @@ class Light(Device):
     @property
     def supports_color(self):
         """Return if light supports color."""
-        return self.color.initialized
+        return self.color.initialized or all(
+            [c.brightness.initialized for c in (self.red, self.green, self.blue)]
+        )
 
     @property
     def supports_rgbw(self):
         """Return if light supports RGBW."""
-        return self.rgbw.initialized
+        return self.rgbw.initialized or all(
+            [
+                c.brightness.initialized
+                for c in (self.red, self.green, self.blue, self.white)
+            ]
+        )
 
     @property
     def supports_tunable_white(self):
@@ -160,6 +281,27 @@ class Light(Device):
     def supports_color_temperature(self):
         """Return if light supports absolute color temperature."""
         return self.color_temperature.initialized
+
+    @classmethod
+    def read_color_from_config(
+        cls, color, config
+    ) -> Tuple[Optional[str], Optional[str], Optional[str], Optional[str]]:
+        """Load color configuration from configuration structure."""
+        if color in config:
+            sub_config = config[color]
+            group_address_switch = sub_config.get("group_address_switch")
+            group_address_switch_state = sub_config.get("group_address_switch_state")
+            group_address_brightness = sub_config.get("group_address_brightness")
+            group_address_brightness_state = sub_config.get(
+                "group_address_brightness_state"
+            )
+            return (
+                group_address_switch,
+                group_address_switch_state,
+                group_address_brightness,
+                group_address_brightness_state,
+            )
+        return None, None, None, None
 
     @classmethod
     def from_config(cls, xknx, name, config):
@@ -183,6 +325,31 @@ class Light(Device):
         min_kelvin = config.get("min_kelvin", Light.DEFAULT_MIN_KELVIN)
         max_kelvin = config.get("max_kelvin", Light.DEFAULT_MAX_KELVIN)
 
+        (
+            red_switch,
+            red_switch_state,
+            red_brightness,
+            red_brightness_state,
+        ) = cls.read_color_from_config("red", config)
+        (
+            green_switch,
+            green_switch_state,
+            green_brightness,
+            green_brightness_state,
+        ) = cls.read_color_from_config("green", config)
+        (
+            blue_switch,
+            blue_switch_state,
+            blue_brightness,
+            blue_brightness_state,
+        ) = cls.read_color_from_config("blue", config)
+        (
+            white_switch,
+            white_switch_state,
+            white_brightness,
+            white_brightness_state,
+        ) = cls.read_color_from_config("white", config)
+
         return cls(
             xknx,
             name,
@@ -198,6 +365,22 @@ class Light(Device):
             group_address_tunable_white_state=group_address_tunable_white_state,
             group_address_color_temperature=group_address_color_temperature,
             group_address_color_temperature_state=group_address_color_temperature_state,
+            group_address_switch_red=red_switch,
+            group_address_switch_state_red=red_switch_state,
+            group_address_brightness_red=red_brightness,
+            group_address_brightness_state_red=red_brightness_state,
+            group_address_switch_green=green_switch,
+            group_address_switch_state_green=green_switch_state,
+            group_address_brightness_green=green_brightness,
+            group_address_brightness_state_green=green_brightness_state,
+            group_address_switch_blue=blue_switch,
+            group_address_switch_state_blue=blue_switch_state,
+            group_address_brightness_blue=blue_brightness,
+            group_address_brightness_state_blue=blue_brightness_state,
+            group_address_switch_white=white_switch,
+            group_address_switch_state_white=white_switch_state,
+            group_address_brightness_white=white_brightness,
+            group_address_brightness_state_white=white_brightness_state,
             min_kelvin=min_kelvin,
             max_kelvin=max_kelvin,
         )
@@ -232,7 +415,51 @@ class Light(Device):
             )
         )
 
-        return '<Light name="{}" ' 'switch="{}"{}{}{}{}{} />'.format(
+        str_red_state = (
+            ""
+            if not self.red.switch.initialized
+            else f' red_state="{self.red.switch.group_addr_str()}"'
+        )
+        str_red_brightness = (
+            ""
+            if not self.red.brightness.initialized
+            else f' red_brightness="{self.red.brightness.group_addr_str()}"'
+        )
+
+        str_green_state = (
+            ""
+            if not self.green.switch.initialized
+            else f' green_state="{self.green.switch.group_addr_str()}"'
+        )
+        str_green_brightness = (
+            ""
+            if not self.green.brightness.initialized
+            else f' green_brightness="{self.green.brightness.group_addr_str()}"'
+        )
+
+        str_blue_state = (
+            ""
+            if not self.blue.switch.initialized
+            else f' blue_state="{self.blue.switch.group_addr_str()}"'
+        )
+        str_blue_brightness = (
+            ""
+            if not self.blue.brightness.initialized
+            else f' blue_brightness="{self.blue.brightness.group_addr_str()}"'
+        )
+
+        str_white_state = (
+            ""
+            if not self.white.switch.initialized
+            else f' white_state="{self.white.switch.group_addr_str()}"'
+        )
+        str_white_brightness = (
+            ""
+            if not self.white.brightness.initialized
+            else f' white_brightness="{self.white.brightness.group_addr_str()}"'
+        )
+
+        return '<Light name="{}" ' 'switch="{}"{}{}{}{}{}{}{}{}{}{}{}{}{} />'.format(
             self.name,
             self.switch.group_addr_str(),
             str_brightness,
@@ -240,20 +467,43 @@ class Light(Device):
             str_rgbw,
             str_tunable_white,
             str_color_temperature,
+            str_red_state,
+            str_red_brightness,
+            str_green_state,
+            str_green_brightness,
+            str_blue_state,
+            str_blue_brightness,
+            str_white_state,
+            str_white_brightness,
         )
 
     @property
     def state(self) -> Optional[bool]:
         """Return the current switch state of the device."""
-        return self.switch.value
+        if self.switch.value is not None:
+            return self.switch.value
+        if any(
+            [
+                c.switch.value is not None
+                for c in (self.red, self.green, self.blue, self.white)
+            ]
+        ):
+            return any(
+                [c.switch.value for c in (self.red, self.green, self.blue, self.white)]
+            )
+        return None
 
     async def set_on(self):
         """Switch light on."""
         await self.switch.on()
+        for color in (self.red, self.green, self.blue, self.white):
+            await color.set_on()
 
     async def set_off(self):
         """Switch light off."""
         await self.switch.off()
+        for color in (self.red, self.green, self.blue, self.white):
+            await color.set_off()
 
     @property
     def current_brightness(self):
@@ -275,10 +525,21 @@ class Light(Device):
         If the device supports RGBW, get the current RGB+White values instead.
         """
         if self.supports_rgbw:
-            if not self.rgbw.value:
-                return None, None
-            return self.rgbw.value[:3], self.rgbw.value[3]
-        return self.color.value, None
+            if self.rgbw.initialized:
+                if not self.rgbw.value:
+                    return None, None
+                return self.rgbw.value[:3], self.rgbw.value[3]
+        if self.color.initialized:
+            return self.color.value, None
+        # individual RGB addresses - white will return None when it is not initialized
+        colors = [
+            self.red.brightness.value,
+            self.green.brightness.value,
+            self.blue.brightness.value,
+        ]
+        if None in colors:
+            return None, self.white.brightness.value
+        return colors, self.white.brightness.value
 
     async def set_color(self, color, white=None):
         """
@@ -289,13 +550,36 @@ class Light(Device):
         """
         if white is not None:
             if self.supports_rgbw:
-                await self.rgbw.set(list(color) + [white])
-                return
+                if self.rgbw.initialized:
+                    await self.rgbw.set(list(color) + [white])
+                    return
+                if all(
+                    [
+                        c.brightness.initialized
+                        for c in (self.red, self.green, self.blue, self.white)
+                    ]
+                ):
+                    await self.red.brightness.set(color[0])
+                    await self.green.brightness.set(color[1])
+                    await self.blue.brightness.set(color[2])
+                    await self.white.brightness.set(white)
+                    return
             logger.warning("RGBW not supported for device %s", self.get_name())
         else:
             if self.supports_color:
-                await self.color.set(color)
-                return
+                if self.color.initialized:
+                    await self.color.set(color)
+                    return
+                if all(
+                    [
+                        c.brightness.initialized
+                        for c in (self.red, self.green, self.blue)
+                    ]
+                ):
+                    await self.red.brightness.set(color[0])
+                    await self.green.brightness.set(color[1])
+                    await self.blue.brightness.set(color[2])
+                    return
             logger.warning("Colors not supported for device %s", self.get_name())
 
     @property
